+++ client/src/commonMain/kotlin/org/apptank/horus/client/websocket/WebSocketSyncEventsSubscriber.kt
@@ -1,0 +1,154 @@
+package org.apptank.horus.client.websocket
+
+import io.ktor.client.HttpClient
+import io.ktor.client.plugins.websocket.webSocket
+import io.ktor.websocket.Frame
+import io.ktor.websocket.readText
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.channels.Channel
+import kotlinx.coroutines.currentCoroutineContext
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.isActive
+import kotlinx.coroutines.launch
+import kotlinx.serialization.encodeToString
+import kotlinx.serialization.json.Json
+import org.apptank.horus.client.base.coFold
+import org.apptank.horus.client.config.HorusConfig
+import org.apptank.horus.client.control.SyncControl
+import org.apptank.horus.client.extensions.info
+import org.apptank.horus.client.extensions.logException
+import org.apptank.horus.client.sync.network.service.IBroadcastService
+import org.apptank.horus.client.sync.network.dto.SyncDTO
+import org.apptank.horus.client.sync.network.dto.toDomain
+import org.apptank.horus.client.websocket.data.SocketConnectionData
+import org.apptank.horus.client.websocket.data.SubscribeData
+import org.apptank.horus.client.websocket.data.WebSocketEvent
+import org.apptank.horus.client.websocket.data.WebSocketPusherEventName
+import kotlin.math.min
+
+class WebSocketSyncEventsSubscriber(
+    private val httpClient: HttpClient,
+    private val config: HorusConfig,
+    private val broadcastService: IBroadcastService
+) {
+    private val decoderJson = Json {
+        ignoreUnknownKeys = true
+        encodeDefaults = true
+    }
+    internal val outgoingChannel = Channel<WebSocketEvent>(Channel.BUFFERED)
+
+    suspend fun subscriber(ownerId: String, onActionReceived: (SyncControl.Action) -> Unit) {
+
+        var currentDelay = 1000L
+
+        config.websocketConfig ?: return
+
+        // ---------------------------------------
+        // LIFECYCLE SOCKET
+        // ---------------------------------------
+
+        while (currentCoroutineContext().isActive) {
+
+            var sendJob: Job? = null
+
+            try {
+                val websocketUrl = "${config.websocketConfig.baseUrl}/horus/${config.websocketConfig.authKey}"
+                info("[WebSocketSyncEventsSubscriber] Starting WebSocket connection: ${websocketUrl}")
+
+                httpClient.webSocket(urlString = websocketUrl) {
+                    currentDelay = 1000L
+
+                    // -------------------------------------
+                    // OUTPUT DATA
+                    // -------------------------------------
+
+                    sendJob = launch {
+                        for (event in outgoingChannel) {
+                            val jsonText = decoderJson.encodeToString(event)
+                            info("[WebSocketSyncEventsSubscriber] Sending event: $jsonText")
+                            send(Frame.Text(jsonText))
+                        }
+                    }
+
+                    // -------------------------------------
+                    // INCOMING DATA
+                    // -------------------------------------
+
+                    for (frame in incoming) {
+                        processFrame(frame, ownerId, onActionReceived)
+                    }
+                }
+            } catch (e: Exception) {
+                sendJob?.cancel()
+                logException("[WebSocketSyncEventsSubscriber] Error while receiving WebSocket events", e)
+                if (!currentCoroutineContext().isActive) break
+            }
+
+            // RECONNECTION BACKOFF
+            delay(currentDelay)
+            currentDelay = min(currentDelay * 2, MAX_DELAY)
+        }
+    }
+
+    internal suspend fun processFrame(
+        frame: Frame,
+        ownerId: String,
+        onActionReceived: (SyncControl.Action) -> Unit
+    ) {
+        if (frame is Frame.Text) {
+            val text = frame.readText()
+
+            when {
+                text.contains(WebSocketPusherEventName.PING.id) -> {
+                    outgoingChannel.send(
+                        WebSocketEvent(WebSocketPusherEventName.PONG.id)
+                    )
+                }
+
+                text.contains(WebSocketPusherEventName.CONNECTION_ESTABLISHED.id) -> {
+                    setupSubscriberChannel(ownerId, text)
+                }
+
+                text.contains(WebSocketPusherEventName.SUBSCRIPTION_SUCCEEDED.id) -> {
+                    info("[WebSocketSyncEventsSubscriber] Subscription succeeded")
+                }
+
+                text.contains(WebSocketPusherEventName.SYNC_ACTION.id) -> {
+                    val webSocketEvent = decoderJson.decodeFromString<WebSocketEvent>(text)
+                    info("[WebSocketSyncEventsSubscriber] Received Sync Action: $text")
+                    onActionReceived(decoderJson.decodeFromString<SyncDTO.Response.SyncAction>(webSocketEvent.data).toDomain())
+                }
+
+                else -> {
+                    info("[WebSocketSyncEventsSubscriber] Unknown event received: $text")
+                }
+            }
+        }
+    }
+
+    internal suspend fun setupSubscriberChannel(userOwnerId: String, data: String) {
+        val webSocketEvent = decoderJson.decodeFromString<WebSocketEvent>(data)
+        val socketConnectionData = decoderJson.decodeFromString<SocketConnectionData>(webSocketEvent.data)
+        val channelName = "private-horus.sync.$userOwnerId"
+        info("[WebSocketSyncEventsSubscriber] Channel Authentication...")
+
+        broadcastService.postAuth(socketConnectionData.socketId, "private-horus.sync.$userOwnerId").coFold(
+            onSuccess = {
+                info("[WebSocketSyncEventsSubscriber] Channel Authentication successful: ${it.auth}")
+                outgoingChannel.send(
+                    WebSocketEvent(
+                        WebSocketPusherEventName.SUBSCRIBE.id,
+                        decoderJson.encodeToString(SubscribeData(it.auth, channelName))
+                    )
+                )
+            },
+            onFailure = {
+                logException("[WebSocketSyncEventsSubscriber] Error while setting up subscriber channel", it)
+            }
+        )
+    }
+
+    companion object {
+        const val MAX_DELAY = 30000L
+    }
+}